requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
SQLAlchemy==2.0.30
alembic==1.13.1
redis==5.0.4
aioredis==2.0.1 
pydantic==2.7.1
pydantic-settings==2.6.0
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.27.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
